pyproject.toml
[project]
name = "stammdatenmanagement"
version = "1.0.0"
description = "Stammdatenmanagement - Dubletten-Bereinigung"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.9.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "slowapi>=0.1.9",
    "supabase>=2.0.0",
    "openpyxl>=3.1.0",
]
